Moves the TypeChecker implementation from expr to il/evaluator.
