+++ mixer/pkg/expr/expr_test.go
@@ -229,95 +229,6 @@ func newAF(ds []*ad) *af {
 	return &af{m}
 }
 
-func TestTypeCheck(t *testing.T) {
-	af := newAF([]*ad{
-		{"int", dpb.INT64},
-		{"bool", dpb.BOOL},
-		{"double", dpb.DOUBLE},
-		{"string", dpb.STRING},
-		{"timestamp", dpb.TIMESTAMP},
-		{"ip", dpb.IP_ADDRESS},
-		{"email", dpb.EMAIL_ADDRESS},
-		{"uri", dpb.URI},
-		{"dns", dpb.DNS_NAME},
-		{"duration", dpb.DURATION},
-		{"stringmap", dpb.STRING_MAP},
-	})
-
-	tests := []struct {
-		in  string
-		out dpb.ValueType
-		err string
-	}{
-		// identity
-		{"int", dpb.INT64, ""},
-		{"bool", dpb.BOOL, ""},
-		{"double", dpb.DOUBLE, ""},
-		{"string", dpb.STRING, ""},
-		{"timestamp", dpb.TIMESTAMP, ""},
-		{"ip", dpb.IP_ADDRESS, ""},
-		{"email", dpb.EMAIL_ADDRESS, ""},
-		{"uri", dpb.URI, ""},
-		{"dns", dpb.DNS_NAME, ""},
-		{"duration", dpb.DURATION, ""},
-		{"stringmap", dpb.STRING_MAP, ""},
-		// expressions
-		{"int == 2", dpb.BOOL, ""},
-		{"double == 2.0", dpb.BOOL, ""},
-		{`string | "foobar"`, dpb.STRING, ""},
-		// invalid expressions
-		{"int | bool", dpb.VALUE_TYPE_UNSPECIFIED, "typeError"},
-		{"stringmap | ", dpb.VALUE_TYPE_UNSPECIFIED, "failed to parse"},
-	}
-
-	for idx, tt := range tests {
-		t.Run(fmt.Sprintf("[%d] %s", idx, tt.in), func(t *testing.T) {
-			ev := NewTypeChecker([]FunctionMetadata{})
-			vt, err := ev.EvalType(tt.in, af)
-			if tt.err != "" || err != nil {
-				if !strings.Contains(err.Error(), tt.err) {
-					t.Fatalf("EvalType(%s, adf) = %v, wanted err %v", tt.in, err, tt.err)
-				}
-			}
-			if vt != tt.out {
-				t.Fatalf("EvalType(%s, adf) = %v, wanted type %v", tt.in, vt, tt.out)
-			}
-		})
-	}
-}
-
-func TestAssertType(t *testing.T) {
-	af := newAF([]*ad{
-		{"int64", dpb.INT64},
-		{"string", dpb.STRING},
-		{"duration", dpb.DURATION},
-	})
-
-	tests := []struct {
-		name     string
-		expr     string
-		expected dpb.ValueType
-		err      string
-	}{
-		{"correct type", "string", dpb.STRING, ""},
-		{"wrong type", "int64", dpb.STRING, "expected type STRING"},
-		{"eval error", "duration |", dpb.DURATION, "failed to parse"},
-	}
-
-	for idx, tt := range tests {
-		t.Run(fmt.Sprintf("[%d] %s", idx, tt.name), func(t *testing.T) {
-			ev := NewTypeChecker([]FunctionMetadata{})
-			if err := ev.AssertType(tt.expr, af, tt.expected); tt.err != "" || err != nil {
-				if tt.err == "" {
-					t.Fatalf("AssertType(%s, af, %v) = %v, wanted no err", tt.expr, tt.expected, err)
-				} else if !strings.Contains(err.Error(), tt.err) {
-					t.Fatalf("AssertType(%s, af, %v) = %v, wanted err %v", tt.expr, tt.expected, err, tt.err)
-				}
-			}
-		})
-	}
-}
-
 func TestInternalTypeCheck(t *testing.T) {
 	success := "__SUCCESS__"
 	tests := []struct {
